backend/requirements.txt
Django>=4.1,<5.0
djangorestframework
gunicorn
djongo
redis
loguru
pytest
